Implement bookmark processor
