+++ Cargo.toml
@@ -40,6 +40,7 @@ html2md = "0.2.14"
 readability = "0.2.0"
 async-trait = "0.1.73"
 scraper = "0.18.1"
+parking_lot = "0.12.1"
 
 [dev-dependencies]
 assert_cmd = "2.0.12"
