+++ src/cmd/mod.rs
@@ -11,7 +11,7 @@ mod update;
 pub use add::add;
 pub use clean::clean;
 pub use configure::configure;
-pub use fetch::{fetch, fetch_diff, process_bookmarks};
+pub use fetch::{fetch, fetch_diff};
 pub use import::import;
 pub use init::init;
 pub use remove::remove;
